move location to the `contexts` dir
